code/requirements.txt
kafka-python==2.0.1
pyspark==2.4.6
pyspark-stubs==2.4.0